Destroy connection on webrtc close event
This appears to increase stability of the network 100 fold. I think it's
a bit of an illusion created with the testing environment, though.
What's happening is that previously, when a node lost a connection, _it
didn't clean it up until garbage collection_. This means if that node
immediately hopped back on, it'd take a few seconds for a neighbor to
find it, since the neighbor thought they were already connected. In real
life nodes would probably not just be spontaneously restarting, but hey,
at least the network can handle it gracefully if they do!