Give garbageCollectConnections() a pass
In addition to being cleaner, now it intelligently chooses which
duplicate clientId connection to clean by whether it has channelName or
not. This has increased stability a little bit on startup time.
Connection management is still not perfect though - the race condition
that results in duplicates still needs to be worked out.