Fix race condition in connecting; remove uuid()
Don't need uuid so that's gone.

Introduced is the idea of a PendingConnection - one whose sdp
information has not yet come in. With this type, we can get rid
of all the asynchronicity in connection building. This is how we fix the
race condition.

The race condition was that if another node had queued up many offer
broadcasts, then sent them through an intermediary to us, us not having
connected to them yet, we'd make a new connection for each queued
message. This is because it'd take a second or so for a connection to be
generated and put into our pool which we'd check to see if we were
already connected to a certain client. Now that connection creation
is synchronous that doesn't happen any more.

I think this solution works but does deserve to be refactored.
Connection should be a class and emit an event when it has sdp
information at least.