Refactor connections
Honestly, when I wrote this code originally, I was experiencing an
abundance of excitement about the project that was so powerful I
couldn't help but rush through it a bit. Well, as you guessed, I
introduced some tech debt.

This is the first of hopefully a few big refactors to help clean things
up. It works great and is wildly stable, tests passing and all that. But
most importantly, it's more beautiful and easy to grok.