Iterate on tests
There's a couple tests now and they're in a format that allows for more
to be made.

This was no small feat, it took days to accomplish. The tests are
finicky, and it's hard to tear down one of these networks. Writing
these, running c.peer.destroy() was causing a SIGSEGV segfault in node.
I don't really understand fully what all is going on, all I know is I'm
tired of working on this specific piece and am stoked to have something
that's working at least pretty well.