Fix tests after recent exporting error; amend readme
So changing the way this is exported changed the way it was required as
well, of course. Didn't occur to me later. Oh well! Probably will have
to update database as well. The whole reason for the switch was so we
didn't have to require it like Network.Network in UMD builds. Now it's
just Network.default lol.