Require values for emit/on
Building another app using this library, the other app had a type error
when doing network.on('message', { appId, message } => {...}), that
appId and message are possibly undefined. This aligns with the types in
the code here. However, interestingly, dbdb used it in the same way
without any issues, I do not know what's going on there.