Massive change, will precede the release of next minor version
This change completely reworks the way connections are handled in every
way.

Now there's a new connection for every new address, no more concept of a
single open connection that we broadcast the sdp info for. Connection
itself is more obvious. Both messaging to achieve a connection and
switchboard connections are vastly more robust and fast.

The biggest change here is that the initial connection used to take
10-15 seconds. Now? Now it's usually less than 1!

There's a kind of switchboard backoff in that once a node is connected
to another, it slows down hitting the switchboard.

The switchboard also got a complete rewrite.