Fix a connection issue:
Refreshes with the same address could not reconnect. The switchboard was
getting inundated by hundreds of negotiations just between two nodes.
Things were not good!

Then I noticed we weren't making use of the connectionId field. That did
it!

Now there's a test for it and everything connections so lickety split
it's almost mind numbing. Happy days!