Add encryption to SDP exchange
This changes the way the network passes around SDP info. Previously it
was stored raw. This means everyone on the network could see everyone
elses IP information. Also it means anyone interacting with the
switchboard could see everyone's IP information. This is not so secure,
so now the IP information is exposed only to the node it's meant for via
some nice ECIES from browser-network/crypto.

This change involves changing the 'destroy-connection' event from
passing just the connection id to passing the whole connection, so it
warrants a minor version bump.